requirements.txt
aiohttp
aiohttp_jinja2
aiomysql
pyyaml
